backend/requirements.txt
fastapi
uvicorn
duckdb
bcrypt
python-jose[cryptography]
python-dotenv
